Remove modlists Now has basic ability to apply/remove modlists.

+++ skyrimModManager.sh
@@ -11,16 +11,38 @@ export GUM_CONFIRM_SELECTED_BACKGROUND="4"
 export GUM_CONFIRM_SELECTED_BOLD="true"
 export GUM_SPIN_SPINNER="line"
 
+func_remove_modlist () {
+	a=$(pwd)
+	cd $skyrimDirectory/Data/
+	find [a-z]* -maxdepth 0
+	$a/gum/gum confirm "Remove these files?" && rm -rf $(find [a-z]* -maxdepth 0) && $a/gum/gum style "All modlists have been removed." --foreground="11" || $a/gum/gum style "Modlist removal was canceled." --foreground="11"
+	cd $a
+	func_main_menu
+}
+
 func_apply_modlist_menu () {
-	modlistToApply=$(./gum/gum choose $(ls ./modlists | grep -v FORMAT.md))
+	modlistToApply=$(./gum/gum choose $(ls ./modlists | grep -v FORMAT.md) "=Back=")
+	if [[ $modlistToApply == "=Back=" ]]
+		then
+		func_modlist_options_menu
+		exit
+	fi
+	if [[ $(ls ./modlists | grep -w $modlistToApply) == "" ]]
+		then
+		printf "Error, selected item not in directory.\n"
+		exit
+	fi	
+
 	mkdir ./modlists/temp
 	cp -r ./modlists/$modlistToApply/* modlists/temp/
 	cd ./modlists/temp
+	# Replaces spaces with underscores, prefixes "tagged_" to every file (not directory)
+	for i in *\ *; do mv "$i" "${i// /_}"; done
 	for i in $(ls -p | grep -v /); do mv "$i" "tagged_$i"; done
 	cd ../..
 	rsync -a ./modlists/temp/* $skyrimDirectory/Data/
 	rm -rf ./modlists/temp
-	./gum/gum style "Modlist $modlistToApply has been applied." --foreground="8"
+	./gum/gum style "Modlist $modlistToApply has been applied." --foreground="11"
 	./gum/gum spin --title="" --spinner="dot" sleep 4
 	func_main_menu
 }
@@ -30,14 +52,14 @@ func_launch_game () {
 }
 
 func_modlist_options_menu () {
-	case $(./gum/gum choose "Apply new modlist(s)" "Remove current modlist" "Back to main menu" --header="Modlists" --cursor=" >> ") in 
-		"Apply new modlist(s)")
+	case $(./gum/gum choose "Apply new modlist" "Remove current modlist(s)" "Back" --header="Modlists" --cursor=" >> ") in 
+		"Apply new modlist")
 			func_apply_modlist_menu
 			;;
 		"Remove current modlist(s)")
-			printf "Removing modlist"
+			func_remove_modlist
 			;;
-		"Back to main menu")
+		"Back")
 			func_main_menu
 			;;
 	esac
